Safari-609.1.11.1/LayoutTests/http/tests/workers/service/page-cache-service-worker-pending-promise.https-expected.txt

Tests that a page with a pending service worker promise is able to enter the page cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page was restored from Page Cache
PASS Unregistered successfuly
PASS !!restoredFromPageCache is true
PASS successfullyParsed is true

TEST COMPLETE
